explode everything
